# specs/001-pokemon-card-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ArchiveDex API
  version: 1.0.0
  description: Same-origin API for card capture, review, collection browsing, and specimen management.
  license:
    name: Private use only
    identifier: LicenseRef-Private
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /specimens/{specimenId}/valuation:
    post:
      summary: Schedule a manual valuation refresh
      operationId: refreshValuation
      parameters:
        - name: specimenId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '202':
          description: Valuation refresh scheduled
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /specimens/{specimenId}/valuation-history:
    get:
      summary: Read every recorded valuation attempt for a specimen, newest first
      operationId: getValuationHistory
      parameters:
        - name: specimenId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Valuation history
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ValuationHistoryEntry' }
        '404': { $ref: '#/components/responses/Problem' }
  /specimens/{specimenId}/valuation/review:
    post:
      summary: Accept or reject a valuation the plausibility guard held back
      operationId: resolveValuationReview
      parameters:
        - name: specimenId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [accept, reject] }
      responses:
        '204': { description: Proposal resolved }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /valuations/refresh-all:
    post:
      summary: Start or return the active collection-wide valuation refresh
      operationId: startValuationRefreshAll
      responses:
        '202':
          description: Persisted refresh job accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValuationRefreshJob' }
        '503': { $ref: '#/components/responses/Problem' }
    get:
      summary: Read the latest collection-wide valuation refresh
      operationId: getLatestValuationRefresh
      responses:
        '200':
          description: Latest refresh job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValuationRefreshJob' }
        '204': { description: No refresh has been started }
  /valuations/refresh-all/{jobId}:
    get:
      summary: Read one collection-wide valuation refresh
      operationId: getValuationRefresh
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Refresh job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValuationRefreshJob' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-ArchiveDex.Auth
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: Actionable RFC Problem Details response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        traceId: { type: string }
        retryable: { type: boolean }
        nextAction: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    ValuationRefreshJob:
      type: object
      required: [id, status, totalCards, processedCards, updatedCards, unavailableCards, failedCards, heldCards, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [pending, running, completed, completedWithErrors, failed]
        totalCards: { type: integer, minimum: 0 }
        processedCards: { type: integer, minimum: 0 }
        updatedCards: { type: integer, minimum: 0 }
        unavailableCards: { type: integer, minimum: 0 }
        failedCards: { type: integer, minimum: 0 }
        heldCards: { type: integer, minimum: 0, description: Cards whose new value the guard held back for review. Prices come from the uploaded Cardmarket files where available, and fall back to the configured AI provider otherwise }
        lastError: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ValuationHistoryEntry:
      type: object
      required: [id, amountMinor, currency, provider, method, outcome, valuedAt, marketDataAsOf, recordedAt]
      properties:
        id: { type: string, format: uuid }
        amountMinor: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, enum: [EUR] }
        previousAmountMinor: { type: [integer, 'null'], format: int64, minimum: 0 }
        provider: { type: string }
        method: { type: string }
        confidence: { type: [string, 'null'] }
        conditionApplied: { type: [boolean, 'null'] }
        outcome: { type: string, enum: [accepted, heldForReview, rejected] }
        holdReason: { type: [string, 'null'] }
        sourceUrls: { type: array, items: { type: string, format: uri } }
        valuedAt: { type: string, format: date-time }
        marketDataAsOf: { type: string, format: date-time }
        recordedAt: { type: string, format: date-time }
...